src/chains/extractor.py
import os
import base64
from dotenv import load_dotenv
from langchain_google_genai import ChatGoogleGenerativeAI
from src.schema.invoice import Invoice

load_dotenv()

google_api_key = os.getenv("GOOGLE_API_KEY")

llm = ChatGoogleGenerativeAI(
    model="gemini-2.5-flash-lite",
    temperature=0.0,
    api_key=google_api_key,
)

structured_llm = llm.with_structured_output(Invoice)