requirements.txt
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
pymupdf>=1.24.0
pydantic>=2.0.0
